+++ packages/jest-circus/src/eventHandler.ts
@@ -14,7 +14,7 @@ import {TEST_TIMEOUT_SYMBOL} from './types';
 import {
   addErrorToEachTestUnderDescribe,
   describeBlockHasTests,
-  getTestDuration,
+  getChildDuration,
   invariant,
   makeDescribe,
   makeTest,
@@ -27,6 +27,7 @@ const eventHandler: Circus.EventHandler = (event, state) => {
       break;
     }
     case 'hook_start': {
+      event.hook.startedAt = Date.now();
       event.hook.seenDone = false;
       break;
     }
@@ -46,6 +47,7 @@ const eventHandler: Circus.EventHandler = (event, state) => {
       const describeBlock = makeDescribe(blockName, currentDescribeBlock, mode);
       currentDescribeBlock.children.push(describeBlock);
       state.currentDescribeBlock = describeBlock;
+      state.timings.definitionStart = Date.now();
       break;
     }
     case 'finish_describe_definition': {
@@ -87,6 +89,7 @@ const eventHandler: Circus.EventHandler = (event, state) => {
       if (currentDescribeBlock.parent) {
         state.currentDescribeBlock = currentDescribeBlock.parent;
       }
+      state.timings.definitionEnd = Date.now();
       break;
     }
     case 'add_hook': {
@@ -155,6 +158,10 @@ const eventHandler: Circus.EventHandler = (event, state) => {
       currentDescribeBlock.tests.push(test);
       break;
     }
+    case 'hook_success': {
+      event.hook.duration = getChildDuration(event.hook);
+      break;
+    }
     case 'hook_failure': {
       const {test, describeBlock, error, hook} = event;
       const {asyncError, type} = hook;
@@ -170,6 +177,7 @@ const eventHandler: Circus.EventHandler = (event, state) => {
         invariant(test, 'always present for `*Each` hooks');
         test.errors.push([error, asyncError]);
       }
+      hook.duration = getChildDuration(hook);
       break;
     }
     case 'test_skip': {
@@ -181,7 +189,7 @@ const eventHandler: Circus.EventHandler = (event, state) => {
       break;
     }
     case 'test_done': {
-      event.test.duration = getTestDuration(event.test);
+      event.test.duration = getChildDuration(event.test);
       event.test.status = 'done';
       state.currentlyRunningTest = null;
       break;
@@ -220,6 +228,7 @@ const eventHandler: Circus.EventHandler = (event, state) => {
       break;
     }
     case 'setup': {
+      state.timings.setupStart = Date.now();
       // Uncaught exception handlers should be defined on the parent process
       // object. If defined on the VM's process object they just no op and let
       // the parent process crash. It might make sense to return a `dispatch`